% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{asignar_atributos}
\alias{asignar_atributos}
\title{Asignar los atributos para una variable de acuerdo a una lista}
\usage{
asignar_atributos(x, atributos)
}
\arguments{
\item{x}{Un objeto al que se le asignarán atributos.}

\item{atributos}{Lista nombrada con los atributos a asignar al objeto \code{x}. Cada elemento de la lista debe tener un nombre que será el nombre del atributo.}
}
\value{
El objeto \code{x} con los atributos asignados.
}
\description{
Ocupando la matriz de perfilamiento se asignan los atributos a los vectores que componen un data frame
}
