% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{asignar_clase}
\alias{asignar_clase}
\title{Asigna clases a columnas de un data frame}
\usage{
asignar_clase(data, clases)
}
\arguments{
\item{data}{Un data frame cuyas columnas se desean reclasificar.}

\item{clases}{Un vector de caracteres indicando la clase a asignar a cada columna de \code{data}.}
}
\value{
Un data frame con las columnas reclasificadas según las clases especificadas.
}
\description{
Esta función toma un data frame y un vector de clases, y asigna la clase correspondiente a cada columna del data frame.
}
\examples{
\dontrun{
library(purrr)
df <- data.frame(a = c("1", "2"), b = c("3.5", "4.5"))
clases <- c("integer", "numeric")
asignar_clase(df, clases)
}
}
