% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\name{f_apply_homologacion}
\alias{f_apply_homologacion}
\title{Aplicar homologación de categorías a un vector}
\usage{
f_apply_homologacion(x, homologacion)
}
\arguments{
\item{x}{Vector de categorías de entrada (puede ser factor, character, etc.).}

\item{homologacion}{Named character vector. Vector de homologación donde los nombres
corresponden a las categorías de entrada y los valores a las categorías de salida.}
}
\value{
Vector con las categorías homologadas.
}
\description{
Esta función aplica un vector de homologación a un vector de categorías de entrada,
mapeando cada categoría a su correspondiente categoría estandarizada de salida.
Si una categoría de entrada no tiene una correspondencia en el vector de homologación,
se asigna la categoría "Sin categoría".
}
