% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\name{f_cat_etiquetas}
\alias{f_cat_etiquetas}
\title{Homologación de categorías de una variable categórica}
\usage{
f_cat_etiquetas(tbl_nombre, var, conexion)
}
\arguments{
\item{tbl_nombre}{Character. Nombre de la tabla de entrada en el esquema de metadatos.}

\item{var}{Character. Nombre de la variable categórica para la cual se requiere homologación.}

\item{conexion}{DBIConnection. Conexión activa a la base de datos
(por ejemplo, generada con \code{DBI::dbConnect} o \code{RPostgres::dbConnect}).}
}
\value{
Named character vector. Vector de homologación donde los nombres corresponden
a las descripciones de las categorías de salida y los valores a los códigos
de las categorías de salida.
}
\description{
Esta función obtiene un vector de homologación entre las categorías de entrada y salida
para una variable categórica específica de una tabla, utilizando una conexión a base de datos.
El vector resultante permite mapear categorías de entrada a sus correspondientes categorías
estandarizadas de salida.
}
\examples{
\dontrun{
# Obtener vector de homologación
homolog <- f_cat_etiquetas("encuesta_hogares", "estado_civil", conexion = mi_con)
}

}
