% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{f_cat_homolog}
\alias{f_cat_homolog}
\title{Homologación de categorías de una variable categórica}
\usage{
f_cat_homolog(tbl_nombre, var, etiqueta = FALSE, conexion)
}
\arguments{
\item{tbl_nombre}{Character. Nombre de la tabla de entrada.}

\item{var}{Character. Nombre de la variable categórica a homologar.}

\item{etiqueta}{Logical. Si es TRUE, utiliza la descripción de la categoría; si es FALSE, utiliza el código de salida. Por defecto es FALSE.}

\item{conexion}{DBIConnection. Conexión activa a la base de datos (por ejemplo, generada con \code{RPostgres::dbConnect}).}
}
\value{
Named character vector. Vector de homologación donde los nombres corresponden a las categorías de entrada y los valores a las categorías homologadas.
}
\description{
Esta función obtiene un vector de homologación entre las categorías de entrada y salida
para una variable categórica específica de una tabla, utilizando una conexión a base de datos.
}
