% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{f_valida_anio}
\alias{f_valida_anio}
\title{Valida y genera la variable año a partir de metadatos del insumo}
\usage{
f_valida_anio(tabla, conexion, insumo)
}
\arguments{
\item{tabla}{data.frame o tibble que contiene los datos a procesar.}

\item{conexion}{Objeto de conexión a base de datos compatible con \code{DBI}.}

\item{insumo}{Identificador del insumo (\code{id_insumo}) utilizado para filtrar la consulta.}
}
\value{
Un objeto de tipo data.frame o tibble con la columna \code{anio} añadida
si corresponde. Se incluye el atributo \code{process_anio} con valores
\code{"yes"} o \code{"no"} indicando si se realizó el proceso.
}
\description{
Esta función verifica si la tabla de entrada contiene una columna llamada
\code{anio}. En caso de no existir, consulta la base de datos para extraer
el año desde la fecha asociada al insumo y lo integra a la tabla.
Además, registra mediante un atributo si el procesamiento del año fue realizado.
}
\details{
Si la tabla de entrada ya contiene una variable llamada \code{anio},
no se realiza ninguna consulta ni modificación estructural.
}
