% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{missing_evaluation.character_raw}
\alias{missing_evaluation.character_raw}
\title{Evaluate and categorize missing values in a character vector}
\usage{
\method{missing_evaluation}{character}(x)
}
\arguments{
\item{x}{A character vector to evaluate.}
}
\value{
A tibble with two columns: \code{original} and \code{modificada}.
}
\description{
This function takes a character vector and evaluates whether the elements correspond to common missing values,
such as empty strings, blank spaces, or the text "NULL". It returns a tibble with two columns:
\describe{
\item{original}{The original value, replacing missing values with \code{NA_character_}.}
\item{modificada}{A descriptive label for the type of missing value detected, or \code{NA_character_} if not applicable.}
}
}
\examples{
\dontrun{
missing_evaluation.character(c("", " ", "NULL", "data"))
 }
}
