% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{missing_evaluation}
\alias{missing_evaluation}
\title{Evaluate Missing Values}
\usage{
missing_evaluation(x)
}
\arguments{
\item{x}{An object to be evaluated for missing values.}
}
\value{
The result of the missing value evaluation. The specific return value depends on the method implemented for the class of \code{x}.
}
\description{
Generic function to evaluate missing values in an object.
}
