% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{missing_treatment}
\alias{missing_treatment}
\title{missing_treatment}
\usage{
missing_treatment(data, clean = FALSE, reporte = FALSE)
}
\arguments{
\item{data}{Un data.frame que contiene los datos a evaluar.}

\item{clean}{Lógico. Si es TRUE, retorna solo las columnas originales sin sufijo "_original". Por defecto es FALSE.}

\item{reporte}{Lógico. Si es TRUE, retorna solo las columnas modificadas sin sufijo "_modificada". Por defecto es FALSE.}
}
\value{
Un data.frame con el tratamiento de valores faltantes, dependiendo de los argumentos \code{clean} y \code{reporte}.
}
\description{
Esta función evalúa y trata los valores faltantes en un data.frame, permitiendo limpiar o generar reportes según los argumentos proporcionados.
}
\details{
Utiliza \code{purrr::imap_dfc}, \code{dplyr::rename_with}, \code{dplyr::select}, \code{dplyr::mutate}, \code{stringr::str_c}, y \code{stringr::str_remove}.
La función interna \code{missing_evaluation} debe estar definida en el entorno.
}
\examples{
\dontrun{
missing_treatment(mi_data)
missing_treatment(mi_data, clean = TRUE)
missing_treatment(mi_data, reporte = TRUE)
}
}
