% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.cadena}
\alias{perfilamiento.cadena}
\title{Perfilamiento de objetos tipo character}
\usage{
\method{perfilamiento}{cadena}(x)
}
\arguments{
\item{x}{Un objeto de tipo character que se desea transformar.}
}
\value{
Un objeto de tipo character transformado según el tipo especificado en la clase.
}
\description{
Esta función evalúa y transforma un objeto de tipo character.
}
