% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.categorica}
\alias{perfilamiento.categorica}
\title{Perfilamiento de variable categórica}
\usage{
\method{perfilamiento}{categorica}(x)
}
\arguments{
\item{x}{Un vector categórico con un atributo \code{diccionario} que mapea los valores de \code{x} a sus descripciones.}
}
\value{
Un vector con las descripciones correspondientes a los valores de \code{x} según el diccionario.
}
\description{
Esta función toma un vector categórico y devuelve su representación según un diccionario asociado como atributo.
}
