% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.decimal}
\alias{perfilamiento.decimal}
\title{Perfilamiento de objetos de tipo numérico
Esta función evalúa y transforma un objeto de tipo numérico.}
\usage{
\method{perfilamiento}{numeric}(x)
}
\arguments{
\item{x}{Un objeto de tipo numérico que se desea transformar.}
}
\value{
Un objeto de tipo numérico transformado según el tipo especificado en la clase.
}
\description{
Perfilamiento de objetos de tipo numérico
Esta función evalúa y transforma un objeto de tipo numérico.
}
