% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.entero}
\alias{perfilamiento.entero}
\title{Perfilamiento de objetos de tipo entero}
\usage{
\method{perfilamiento}{entero}(x)
}
\arguments{
\item{x}{Un objeto de tipo entero que se desea transformar.}
}
\value{
Un objeto de tipo entero transformado según el tipo especificado en la clase.
}
\description{
Esta función evalúa y transforma un objeto de tipo entero.
}
