% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.fecha}
\alias{perfilamiento.fecha}
\title{Perfilamiento de fechas
Esta función evalúa y transforma un objeto de tipo fecha, aplicando un formato específico definido en los atributos del objeto.}
\usage{
\method{perfilamiento}{fecha}(x)
}
\arguments{
\item{x}{Un objeto de tipo fecha que se desea transformar.}
}
\value{
Un objeto de tipo fecha transformado según el tipo especificado en la clase.
}
\description{
Perfilamiento de fechas
Esta función evalúa y transforma un objeto de tipo fecha, aplicando un formato específico definido en los atributos del objeto.
}
