% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.fecha_hora}
\alias{perfilamiento.fecha_hora}
\title{Perfilamiento de objetos de tipo fecha y hora}
\usage{
\method{perfilamiento}{fecha_hora}(x)
}
\arguments{
\item{x}{Un objeto de tipo fecha y hora que se desea transformar.}
}
\value{
Un objeto de tipo fecha y hora transformado según el tipo especificado en la clase.
}
\description{
Esta función evalúa y transforma un objeto de tipo fecha y hora, aplicando un formato específico definido en los atributos del objeto.
}
