% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.logico}
\alias{perfilamiento.logico}
\title{Perfilamiento de objetos de tipo lógico
Esta función evalúa y transforma un objeto de tipo lógico.}
\usage{
\method{perfilamiento}{logico}(x)
}
\arguments{
\item{x}{Un objeto de tipo lógico que se desea transformar.}
}
\value{
Un objeto de tipo lógico transformado según el tipo especificado en la clase.
}
\description{
Perfilamiento de objetos de tipo lógico
Esta función evalúa y transforma un objeto de tipo lógico.
}
