% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento}
\alias{perfilamiento}
\title{Perfilamiento genérico}
\usage{
perfilamiento(x)
}
\arguments{
\item{x}{Un objeto sobre el que se aplicará el método de perfilamiento.}
}
\value{
El resultado dependerá del método específico implementado para la clase de \code{x}.
}
\description{
Esta función genérica define el método \code{perfilamiento} para su uso con diferentes clases de objetos.
}
