% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilamiento.R
\name{perfilamiento.ruc}
\alias{perfilamiento.ruc}
\title{Limpieza de RUCS}
\usage{
\method{perfilamiento}{ruc}(x)
}
\arguments{
\item{x}{Character. Nombre de la tabla de entrada.}
}
\value{
Named character vector. Vector de homologación donde los nombres corresponden a las categorías de entrada y los valores a las categorías homologadas.
}
\description{
Los RUCS deben 13 digitos y terminar 001 y no tener ningun otro caracter
}
