% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacion.R
\name{valida_dpa}
\alias{valida_dpa}
\title{Valida la presencia y consistencia de niveles DPA en una tabla}
\usage{
valida_dpa(tabla, conexion, insumo)
}
\arguments{
\item{tabla}{Un \code{data.frame} o \code{tibble} que contiene las columnas a validar.}

\item{conexion}{Objeto de conexión a base de datos compatible con \code{DBI}.}

\item{insumo}{Identificador del insumo (\code{id_insumo}) utilizado para filtrar la consulta.}
}
\value{
Invisiblemente, \code{NULL}. La función imprime mensajes informativos
sobre la validación de los niveles DPA.
@importFrom DBI dbGetQuery
@importFrom glue glue
@export
}
\description{
Esta función verifica si una tabla contiene las variables de niveles
administrativos (DPA) requeridas según un insumo específico. Se asegura de que
la presencia de niveles inferiores (canton, parroquia) esté acompañada por
los niveles superiores correspondientes (provincia, canton).
Además, proporciona mensajes informativos sobre la presencia o ausencia
de estos niveles en la tabla y en el insumo.
}
