% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologacion.R
\name{asignar_nombres_columnas}
\alias{asignar_nombres_columnas}
\title{Función para asignar nombres finales a las columas de un dataframe}
\usage{
asignar_nombres_columnas(df, nuevos_nombres)
}
\value{
data.frame con los nuevos nombres de columnas asignados
}
\description{
Asignar nuevos nombres a las columnas de un data.frame
}
